refactor(api): adjust resources namespace
