+++ services/api/app/Http/Controllers/V1/FeatureController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\V1;
 
 use App\Http\Controllers\Controller;
-use App\Http\Resources\FeatureResource;
+use App\Http\Resources\V1\FeatureResource;
 use Laravel\Pennant\Feature;
 
 class FeatureController extends Controller
