+++ services/api/app/Http/Controllers/V1/UserController.php
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers\V1;
 
 use App\Http\Controllers\Controller;
-use App\Http\Resources\UserMeResource;
-use App\Http\Resources\UserResource;
+use App\Http\Resources\V1\UserMeResource;
+use App\Http\Resources\V1\UserResource;
 use App\Models\User;
 use Illuminate\Http\Request;
